+++ HomeAssistantTaskerPlugin/src/main/AndroidManifest.xml
@@ -1,0 +1,109 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Because plug-ins interact with the host in the background, plug-ins MUST be installed to
+     internal memory. This is consistent with Google's app install location guidelines:
+     <http://developer.android.com/guide/appendix/install-location.html#Should>.
+-->
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          xmlns:tools="http://schemas.android.com/tools"
+          package="com.markadamson.taskerplugin.homeassistant"
+          android:installLocation="internalOnly">
+
+    <!-- This allows the plug-in to appear on non-touchscreen devices like Google TV.
+    Be sure to test that the app works without a touchscreen. -->
+    <uses-feature
+            android:name="android.hardware.touchscreen"
+            android:required="false"/>
+
+    <!-- Although many of these default to true, some users with modified ROMs have
+         trouble seeing apps unless these attributes are explicitly set. -->
+    <supports-screens
+            android:anyDensity="true"
+            android:largeScreens="true"
+            android:normalScreens="true"
+            android:smallScreens="true"
+            android:xlargeScreens="true" tools:ignore="UnusedAttribute"/>
+
+    <!-- When hardware acceleration is enabled, the RAM memory usage of the app is
+         significantly increased as soon as any UI is displayed. To reduce the memory
+         usage of a plug-in which is primarily a background app, either (1) hardware
+         acceleration should be disabled or (2) the plug-in's BroadcastReceiver (and any
+         applicable Service) should be put into a different process. This example plug-in
+         creates two processes. -->
+    <application
+            android:name="com.markadamson.taskerplugin.homeassistant.PluginApplication"
+            android:allowBackup="true"
+            android:fullBackupContent="true"
+            android:hardwareAccelerated="true"
+            android:icon="@drawable/ic_launcher"
+            android:label="@string/app_name" tools:ignore="UnusedAttribute">
+
+        <activity
+                android:name="com.markadamson.taskerplugin.homeassistant.ui.activity.EditActivity"
+                android:exported="false"
+                android:label="@string/plugin_name"
+                android:theme="@style/Theme.AppCompat"
+                android:uiOptions="splitActionBarWhenNarrow"
+                android:windowSoftInputMode="adjustResize"
+                tools:ignore="UnusedAttribute">
+        </activity>
+        <!--
+             This is the "edit" Activity. Note that the host will reject plug-in
+             Activities for the following reasons:
+                 - Missing "android:label=[...]"
+                 - Missing "android:icon=[...]"
+                 - The Activity isn't exported (e.g. android:exported="false")
+                 - The Activity isn't enabled (e.g. android:enabled="false")
+                 - The Activity requires permissions not available to the host
+        -->
+        <activity-alias
+                android:name="com.markadamson.taskerplugin.homeassistant.ui.activity.PluginActivity"
+                android:exported="true"
+                android:icon="@drawable/ic_plugin"
+                android:label="@string/plugin_name"
+                android:targetActivity="com.markadamson.taskerplugin.homeassistant.ui.activity.EditActivity"
+                tools:ignore="ExportedActivity">
+            <!-- this Intent filter allows the plug-in to be discovered by the host. -->
+            <intent-filter>
+                <action android:name="com.twofortyfouram.locale.intent.action.EDIT_SETTING"/>
+            </intent-filter>
+        </activity-alias>
+
+        <!--
+             If this plug-in does not stand alone (e.g. it is only a plug-in for Locale and does not
+             have a Launcher Activity), including this in the manifest will help when users try to
+             open the app directly from the app store.
+        -->
+        <activity
+                android:name="com.twofortyfouram.locale.sdk.client.ui.activity.InfoActivity"
+                android:exported="true"
+                tools:ignore="ExportedActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN"/>
+
+                <category android:name="android.intent.category.INFO"/>
+            </intent-filter>
+        </activity>
+
+        <!--
+             This is the "fire" BroadcastReceiver. Note that the host will reject plug-in
+             BroadcastReceivers for the following reasons:
+                - The BroadcastReceiver isn't exported (e.g. android:exported="false")
+                - The BroadcastReceiver isn't enabled (e.g. android:enabled="false")
+                - The BroadcastReceiver requires permissions not available to the host
+                - There are multiple BroadcastReceivers for com.twofortyfouram.locale.intent.action.FIRE_SETTING
+        -->
+        <receiver
+                android:name="com.markadamson.taskerplugin.homeassistant.receiver.FireReceiver"
+                android:exported="true"
+                android:process=":background"
+                tools:ignore="ExportedReceiver">
+
+            <!-- this Intent filter allows the plug-in to discovered by the host. -->
+            <intent-filter>
+                <action android:name="com.twofortyfouram.locale.intent.action.FIRE_SETTING"/>
+            </intent-filter>
+        </receiver>
+    </application>
+
+</manifest>
