+++ HomeAssistantTaskerPlugin/src/main/res/drawable-hdpi/ic_plugin.png
