+++ HomeAssistantTaskerPlugin/src/main/res/drawable-ldpi/ic_launcher.png
