+++ HomeAssistantTaskerPlugin/src/main/res/drawable-ldpi/ic_plugin.png
