+++ HomeAssistantTaskerPlugin/src/main/res/drawable-mdpi/ic_plugin.png
