+++ HomeAssistantTaskerPlugin/src/main/res/drawable-xhdpi/ic_plugin.png
