+++ HomeAssistantTaskerPlugin/src/main/res/menu/menu.xml
