+++ HomeAssistantTaskerPlugin/src/main/res/values/strings.xml
