--- toastPluginSettingApp/src/debug/AndroidManifest.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          xmlns:tools="http://schemas.android.com/tools"
-          package="com.twofortyfouram.locale.example.setting.toast">
-
-    <!-- For code coverage -->
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
-
-    <application/>
-</manifest>
